Car Brand Classification Dataset/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.66.0
scikit-image>=0.22.0
opencv-python>=4.8.0
matplotlib>=3.8.0
streamlit>=1.38.0
